Moved save & reset buttons from the template to the tab view
